requirements.txt
pydantic>=2.0.0
langchain-text-splitters>=0.0.1
langchain-community>=0.0.1
pypdf>=3.0.0
chromadb>=0.4.0
langchain-chroma>=0.1.0
langchain-huggingface>=0.0.1
sentence-transformers>=2.2.0
rank_bm25>=0.2.0
fastapi>=0.100.0
uvicorn>=0.20.0
requests>=2.28.0
pandas>=2.0.0
openpyxl>=3.1.0
